+++ src/panels/profile/ha-refresh-tokens-card.js
@@ -96,9 +96,6 @@ class HaRefreshTokens extends LocalizeMixin(EventsMixin(PolymerElement)) {
 
   async _handleDelete(ev) {
     const token = ev.model.item;
-    if (token.is_current) {
-      return;
-    }
     if (
       !(await showConfirmationDialog(this, {
         text: this.localize(
