+++ src/state/connection-mixin.ts
@@ -26,10 +26,6 @@ export const connectionMixin = <T extends Constructor<HassBaseEl>>(
   superClass: T
 ) =>
   class extends superClass {
-    private __hiddenTimeout?: number;
-
-    private __visiblePromiseResolve?: () => void;
-
     protected initializeHass(auth: Auth, conn: Connection) {
       this.hass = {
         auth,
@@ -157,39 +153,6 @@ export const connectionMixin = <T extends Constructor<HassBaseEl>>(
       subscribeFrontendUserData(conn, "core", (userData) =>
         this._updateHass({ userData })
       );
-
-      document.addEventListener(
-        "visibilitychange",
-        () => this.__handleVisibilityChange(),
-        false
-      );
-    }
-
-    private __handleVisibilityChange() {
-      if (document.hidden) {
-        // If the document is hidden, we will prevent reconnects until we are visible again
-        this.hass!.connection.suspendReconnectUntil(
-          new Promise((resolve) => {
-            this.__visiblePromiseResolve = resolve;
-          })
-        );
-        // We close the connection to Home Assistant after being hidden for 5 minutes
-        this.__hiddenTimeout = window.setTimeout(() => {
-          this.__hiddenTimeout = undefined;
-          this.hass!.connection.suspend();
-        }, 300000);
-      } else {
-        // Clear timer to close the connection
-        if (this.__hiddenTimeout) {
-          clearTimeout(this.__hiddenTimeout);
-          this.__hiddenTimeout = undefined;
-        }
-        // Unsuspend the reconnect
-        if (this.__visiblePromiseResolve) {
-          this.__visiblePromiseResolve();
-          this.__visiblePromiseResolve = undefined;
-        }
-      }
     }
 
     protected hassReconnected() {
