+++ src/state/hass-element.ts
@@ -13,6 +13,7 @@ import ThemesMixin from "./themes-mixin";
 import TranslationsMixin from "./translations-mixin";
 import { urlSyncMixin } from "./url-sync-mixin";
 import ZHADialogMixin from "./zha-dialog-mixin";
+import { suspendMixin } from "./suspend-mixin";
 
 const ext = <T extends Constructor>(baseClass: T, mixins): T =>
   mixins.reduceRight((base, mixin) => mixin(base), baseClass);
@@ -25,6 +26,7 @@ export class HassElement extends ext(HassBaseEl, [
   SidebarMixin,
   DisconnectToastMixin,
   connectionMixin,
+  suspendMixin,
   NotificationMixin,
   dialogManagerMixin,
   urlSyncMixin,
